backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
sqlalchemy==2.0.23
openai==1.3.5
langchain==0.0.335
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0
psutil==5.9.6
aiofiles==23.2.1
alembic==1.13.1
psycopg[binary]==3.1.18
sqlmodel==0.0.16
PyPDF2==3.0.1
python-docx==1.1.0

